+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Scroggins Event Changes
 
+## 1.3.0 (2017-06-12)
+
+#### Changed
+
+* Subhead text.
+* Button text for logging in.
+* Added the encore video player from Vimeo.
+
+#### Notes
+* Still need to update the menu on the main site.
+* Need to note that this _will_ use an existing InfusionSoft form.
+
 ## 1.2.0 (2017-05-17)
 
 ### Change